+++ server/libs/infra/src/utils/pagination.util.ts
@@ -1,16 +1,5 @@
+import { Paginated, PaginationOptions } from '@app/domain';
 import { FindOneOptions, ObjectLiteral, Repository } from 'typeorm';
-
-export interface PaginationOptions {
-  take: number;
-  skip?: number;
-}
-
-export interface PaginationResult<T> {
-  items: T[];
-  hasNextPage: boolean;
-}
-
-export type Paginated<T> = Promise<PaginationResult<T>>;
 
 export async function paginate<Entity extends ObjectLiteral>(
   repository: Repository<Entity>,
@@ -28,17 +17,4 @@ export async function paginate<Entity extends ObjectLiteral>(
   items.splice(paginationOptions.take);
 
   return { items, hasNextPage };
-}
-
-export async function* usePagination<T>(
-  pageSize: number,
-  getNextPage: (pagination: PaginationOptions) => Paginated<T>,
-) {
-  let hasNextPage = true;
-
-  for (let skip = 0; hasNextPage; skip += pageSize) {
-    const result = await getNextPage({ take: pageSize, skip });
-    hasNextPage = result.hasNextPage;
-    yield result.items;
-  }
 }
